+++ solvebio/resource/dataset.py
@@ -2,6 +2,7 @@ from ..client import client
 from ..help import open_help
 from ..query import Query
 from ..utils.humanize import naturalsize
+from ..utils.md5sum import md5sum
 
 from .solveobject import convert_to_solve_object
 from .apiresource import CreateableAPIResource, ListableAPIResource, \
@@ -49,6 +50,45 @@ class Dataset(CreateableAPIResource, ListableAPIResource,
             ['name', 'data_type', 'description'],
             headers=['Field', 'Data Type', 'Description'],
             aligns=['left', 'left', 'left'], sort=True)
+
+        return results
+
+    def template(self, **params):
+        if 'template_url' not in self:
+            raise Exception(
+                'Please use Dataset.retrieve({ID}) before retrieving '
+                'a template')
+
+        response = client.get(self.template_url, params)
+        return convert_to_solve_object(response)
+
+    def commits(self, **params):
+        if 'commits_url' not in self:
+            raise Exception(
+                'Please use Dataset.retrieve({ID}) before looking '
+                'up commits')
+
+        response = client.get(self.commits_url, params)
+        results = convert_to_solve_object(response)
+        results.set_tabulate(
+            ['id', 'title', 'description', 'status', 'created_at'],
+            headers=['ID', 'Title', 'Description', 'Status', 'Created'],
+            aligns=['left', 'left', 'left', 'left', 'left'], sort=False)
+
+        return results
+
+    def imports(self, **params):
+        if 'imports_url' not in self:
+            raise Exception(
+                'Please use Dataset.retrieve({ID}) before looking '
+                'up imports')
+
+        response = client.get(self.imports_url, params)
+        results = convert_to_solve_object(response)
+        results.set_tabulate(
+            ['id', 'title', 'description', 'status', 'created_at'],
+            headers=['ID', 'Title', 'Description', 'Status', 'Created'],
+            aligns=['left', 'left', 'left', 'left', 'left'], sort=False)
 
         return results
 
@@ -148,17 +188,17 @@ class Dataset(CreateableAPIResource, ListableAPIResource,
 
             # Validate the MD5 of the downloaded file.
             # Handle's S3's multipart MD5 calculation.
-            md5sum, blocks = export_file.md5sum(
+            md5, blocks = md5sum(
                 multipart_threshold=manifest['multipart_threshold_bytes'],
                 multipart_chunksize=manifest['multipart_chunksize_bytes']
             )
 
-            if md5sum != manifest_file['md5']:
+            if md5 != manifest_file['md5']:
                 print("### Export failed MD5 verification!")
                 print("### -------------------------------")
                 print("### File: {0}".format(export_file.file_name))
                 print("### Expected MD5: {0}".format(manifest_file['md5']))
-                print("### Calculated MD5: {0}".format(md5sum))
+                print("### Calculated MD5: {0}".format(md5))
                 if blocks and manifest_file['multipart_blocks'] != blocks:
                     print("### Multipart block size failed verification")
                     print("### Expected: {0} blocks"